@@ -2,7 +2,7 @@ package com.anegocios.puntoventa;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
 import android.widget.EditText;
 import android.widget.Toast;
@@ -28,7 +28,7 @@ public class AbrirCajaActivity  extends AppCompatActivity {
     private void mostrarAbrirCaja() {
         Utilerias ut = new Utilerias();
         setContentView(R.layout.abrircaja);
-        realm= ut.obtenerInstanciaBD();
+        realm= ut.obtenerInstanciaBD(this);
 
     }
 
